Edit the attribute of parameter.

+++ app/src/main/java/com/yangp/sample/MainActivity.kt
@@ -141,7 +141,7 @@ class MainActivity : AppCompatActivity(), OnColorClickedListener {
             }
 
             override fun onProgressChanged(seekBar: SeekBar?, progress: Int, fromUser: Boolean) {
-                waveView2.setShapePadding(progress)
+                waveView2.setShapePadding(progress.toFloat())
             }
         })
         switch_animation.setOnCheckedChangeListener { _, isChecked ->
